+++ src/features/references/editor/ReferencesTableView.tsx
@@ -2,32 +2,28 @@ import 'ag-grid-community/styles/ag-grid.css';
 import 'ag-grid-community/styles/ag-theme-alpine.css';
 import './ReferenceView.css';
 
-import {
-  ColDef,
-  ColumnApi,
-  ColumnState,
-  GetRowIdParams,
-  ICellRendererParams,
-  NewValueParams,
-  SelectionChangedEvent,
-} from 'ag-grid-community';
+import { ColDef, GetRowIdParams, ICellRendererParams, NewValueParams, SelectionChangedEvent } from 'ag-grid-community';
 import { AgGridReact } from 'ag-grid-react/lib/agGridReact';
 import { useAtomValue } from 'jotai';
 import { useCallback, useMemo, useRef, useState } from 'react';
-import { IconType } from 'react-icons';
 import { VscDesktopDownload, VscKebabVertical, VscNewFile, VscTable, VscTrash } from 'react-icons/vsc';
 
 import { getReferencesAtom } from '../../../atoms/referencesState';
 import { emitEvent } from '../../../events';
-import { cx } from '../../../lib/cx';
+import { autoFocusAndSelect } from '../../../lib/autoFocusAndSelect';
 import { ReferenceItem, ReferenceItemStatus } from '../../../types/ReferenceItem';
 import { ReferencesItemStatusLabel } from '../components/ReferencesItemStatusLabel';
+import { TopActionIcon } from '../components/TopActionIcon';
 import { UploadTipInstructions } from '../components/UploadTipInstructions';
+import { loadColumnsState, resetColumnsState, saveColumnsState } from './columnsState';
+import { authorsFormatter, firstAuthorFormatter } from './formatters';
 
 export function ReferencesTableView() {
   const references = useAtomValue(getReferencesAtom);
   const [quickFilter, setQuickFilter] = useState('');
   const [numberOfSelectedRows, setNumberOfSelectedRows] = useState(0);
+
+  const gridRef = useRef<AgGridReact<ReferenceItem>>(null);
 
   const handleTitleEdit = useCallback(({ oldValue, newValue }: NewValueParams<ReferenceItem, 'string'>) => {
     console.log('EDIT');
@@ -92,8 +88,6 @@ export function ReferencesTableView() {
     [handleTitleEdit],
   );
 
-  const gridRef = useRef<AgGridReact<ReferenceItem>>(null);
-
   return (
     <div className="flex w-full flex-col overflow-y-auto p-6">
       {references.length === 0 && <UploadTipInstructions />}
@@ -109,21 +103,21 @@ export function ReferencesTableView() {
           onInput={(e) => setQuickFilter(e.currentTarget.value)}
         />
         <div className="text-md flex items-center gap-2">
-          <TopActionItem
+          <TopActionIcon
             action="Add"
             icon={VscNewFile}
             onClick={() => emitEvent('refstudio://menu/references/upload')}
           />
           <VscKebabVertical />
-          <TopActionItem
+          <TopActionIcon
             action="Remove"
             disabled={numberOfSelectedRows === 0}
             icon={VscTrash}
             selectedCount={numberOfSelectedRows}
           />
-          <TopActionItem action="Export" disabled icon={VscDesktopDownload} />
+          <TopActionIcon action="Export" disabled icon={VscDesktopDownload} />
           <VscKebabVertical />
-          <TopActionItem
+          <TopActionIcon
             action="Reset Cols"
             icon={VscTable}
             onClick={() => resetColumnsState(gridRef.current!.columnApi)}
@@ -153,80 +147,13 @@ export function ReferencesTableView() {
   );
 }
 
-const COLUMNS_STATE_STORAGE_KEY = 'referencesTableView.cols';
-function loadColumnsState(columnApi: ColumnApi) {
-  const cols = localStorage.getItem(COLUMNS_STATE_STORAGE_KEY);
-  if (cols) {
-    const state = JSON.parse(cols) as ColumnState[];
-    columnApi.applyColumnState({ state, applyOrder: true });
-  }
-}
-
-function saveColumnsState(columnApi: ColumnApi, ignoreSave = false) {
-  if (!ignoreSave) {
-    const cols = columnApi.getColumnState();
-    localStorage.setItem(COLUMNS_STATE_STORAGE_KEY, JSON.stringify(cols));
-  }
-}
-
-function resetColumnsState(columnApi: ColumnApi) {
-  columnApi.resetColumnState();
-  localStorage.removeItem(COLUMNS_STATE_STORAGE_KEY);
-}
-
-function firstAuthorFormatter({ value }: { value: ReferenceItem['authors'] }) {
-  const [author = undefined] = value;
-  return author?.fullName ?? '';
-}
-
-function authorsFormatter({ value }: { value: ReferenceItem['authors'] }) {
-  return value.map((a) => a.fullName.split(' ').pop()).join(', ');
-}
-
 function StatusCell({ value }: ICellRendererParams<ReferenceItem, ReferenceItemStatus>) {
   if (!value) {
     return null;
   }
   return (
-    <span className="text-xs">
+    <span className="text-[10px]">
       <ReferencesItemStatusLabel status={value} />
-    </span>
-  );
-}
-
-function autoFocusAndSelect(input: HTMLInputElement | null) {
-  input?.focus();
-  input?.select();
-}
-
-function TopActionItem({
-  action,
-  icon: Icon,
-  selectedCount = 0,
-  disabled,
-  onClick,
-}: {
-  action: string;
-  selectedCount?: number;
-  disabled?: boolean;
-  icon: IconType;
-  onClick?: () => void;
-}) {
-  return (
-    <span
-      className={cx(
-        'flex items-center gap-1 rounded-md px-2 py-1',
-        'cursor-pointer',
-        'border border-slate-400 hover:bg-slate-400 hover:text-white',
-        {
-          'pointer-events-none bg-slate-100 text-slate-400': disabled,
-        },
-      )}
-      onClick={onClick}
-    >
-      <Icon />
-      {action}
-      {selectedCount > 0 && <span>({selectedCount})</span>}
     </span>
   );
 }
