+++ src/lib/autoFocusAndSelect.tsx
@@ -1,0 +1,4 @@
+export function autoFocusAndSelect(input: HTMLInputElement | null) {
+  input?.focus();
+  input?.select();
+}
